BILl-66 - separate controller for fee entry
