+++ src/routes/index.js
@@ -4,6 +4,10 @@ import {
   showLawCategoryPage,
 } from "../controllers/lawCategoryController";
 import { getSessionData } from "../utils";
+import {
+  postFeeEntryPage,
+  showFeeEntryPage,
+} from "../controllers/feeEntryController";
 export const router = express.Router();
 
 router.get("/", (req, res) => {
@@ -29,46 +33,9 @@ router.get("/law-category", showLawCategoryPage);
 
 router.post("/law-category", postLawCategoryPage);
 
-router.get("/fee-entry", (req, res) => {
-  try {
-    getSessionData(req);
-    res.render("main/feeEntry", { csrfToken: req.csrfToken() });
-  } catch (ex) {
-    console.error("Error loading page /fee-entry: {}", ex.message);
+router.get("/fee-entry", showFeeEntryPage);
 
-    res.render("main/error", {
-      status: "An error occurred",
-      error: "An error occurred.",
-    });
-  }
-});
-
-router.post("/fee-entry", async (req, res) => {
-  try {
-    const fee = req.body.fee;
-
-    if (fee == null || isNaN(fee)) {
-      throw new Error("Fee not defined");
-    }
-
-    const response = await req.axiosMiddleware.post("/fees/" + fee);
-    const number = response.data;
-    console.log(req.session?.data);
-
-    // Save this so it can be displayed on the result page
-    req.session.data.result = number;
-    console.log(req.session?.data);
-
-    res.redirect("/result");
-  } catch (ex) {
-    console.error("Error occurred during POST /fee-entry: {}", ex.message);
-
-    res.render("main/error", {
-      status: "An error occurred",
-      error: "An error occurred posting the answer.",
-    });
-  }
-});
+router.post("/fee-entry", postFeeEntryPage);
 
 router.get("/result", (req, res) => {
   try {
